fix: Cast OTP size env variable to integer
Ensures the 'otp.size' configuration value is always an integer by casting the OTP_SIZE environment variable. This prevents potential type issues if the environment variable is set as a string.

+++ config/auth.php
@@ -117,6 +117,6 @@ return [
     'password_timeout' => env('AUTH_PASSWORD_TIMEOUT', 10800),
 
     'otp' => [
-        'size' => env('OTP_SIZE', 6),
+        'size' => (int) env('OTP_SIZE', 6),
     ],
 ];
